Moved get_intended_text and get_input_text_from_response into Utils
