+++ test/general/test_client_server_interaction.py
@@ -1,6 +1,6 @@
 import unittest
 
-from util.server_communication import get_intended_text, get_input_text_from_response
+from Utils import get_intended_text, get_input_text_from_response
 
 
 class TestClient(unittest.TestCase):
